move code to tf.contrib
