+++ tensorflow/contrib/BUILD
@@ -52,6 +52,7 @@ cc_library(
         "//tensorflow/contrib/layers:sparse_feature_cross_op_kernel",
         "//tensorflow/contrib/linear_optimizer:sdca_op_kernels",
         "//tensorflow/contrib/metrics:set_ops_kernels",
+        "//tensorflow/contrib/framework/kernels:all_kernels",
     ],
 )
 
@@ -64,6 +65,7 @@ cc_library(
         "//tensorflow/contrib/layers:sparse_feature_cross_op_op_lib",
         "//tensorflow/contrib/linear_optimizer:sdca_ops_op_lib",
         "//tensorflow/contrib/metrics:set_ops_op_lib",
+        "//tensorflow/contrib/framework:all_ops",
     ],
 )
