+++ tensorflow/contrib/framework/kernels/zero_initializer_op.cc
@@ -13,8 +13,9 @@ See the License for the specific language governing permissions and
 limitations under the License.
 ==============================================================================*/
 
-#include "tensorflow/core/kernels/zero_initializer_op.h"
+#include "tensorflow/contrib/framework/kernels/zero_initializer_op.h"
 #include "tensorflow/core/framework/register_types.h"
+#include "tensorflow/core/framework/op_kernel.h"
 
 namespace tensorflow {
 #define REGISTER_KERNELS(D, T) \
@@ -22,14 +23,12 @@ namespace tensorflow {
       .Device(DEVICE_##D) \
       .TypeConstraint<T>("T"), \
       ZeroInitializerOp<T>);
-#define REGISTER_CPU_KERNELS(T) \
-  REGISTER_KERNELS(CPU, T);
+#define REGISTER_CPU_KERNELS(T) REGISTER_KERNELS(CPU, T);
 TF_CALL_REAL_NUMBER_TYPES(REGISTER_CPU_KERNELS);
 #undef REGISTER_CPU_KERNELS
 
 #if GOOGLE_CUDA
-#define REGISTER_GPU_KERNELS(T) \
-  REGISTER_KERNELS(GPU, T);
+#define REGISTER_GPU_KERNELS(T) REGISTER_KERNELS(GPU, T);
 TF_CALL_GPU_NUMBER_TYPES(REGISTER_GPU_KERNELS);
 #undef REGISTER_GPU_KERNELS
 #endif // GOOGLE_CUDA
